Add loop function.
uCmd_Loop function allows to monitor line buffer. When a
complete message has been received, the uCmd module executes
on the complete raw message.

A lock macro was added to make sure that there are no race
conditions between the line module, that may be run inside
an char received interupt, and the uCmd module that may be
run inside an RTOS task. The user is responsible of providing
this lock/unlock functionality inside the ucmd_lock.h file.